dialogv: int oldheight → long, fixes LP64 stack-stomp hiding dialog text
FossilOrigin-Name: 467bd2089d5bdd6b123a3e6322c140922f521437d4e8690b226ce2a06b217b9d

+++ src/atk/utils/dialogv.c
@@ -299,7 +299,7 @@ long *dHeight;
 {
     long iw, ih;
     long dummy;
-    int  oldheight;
+    long oldheight;
     struct dialog *d=(struct dialog *)dialogv_GetDataObject(self);
     int style;
